+++ config.php
@@ -195,4 +195,4 @@ $THEME->scss = function(theme_config $theme) {
 $THEME->extrascsscallback = 'theme_foundation_extra_scss';
 
 $THEME->iconsystem = '\\theme_foundation\\output\\icon_system_fontawesome';
-$THEME->haseditswitch = true;
+$THEME->haseditswitch = \theme_foundation\toolbox::get_config_setting('navbareditswitch');