+++ Changes.md
@@ -2,7 +2,7 @@ Version Information
 ===================
 Version 4.0.0.1 - TBR
 ---------------------------
-1. Add core edit switch.
+1. Add core edit switch with 'navbareditswitch' setting to enable / disable.
 2. Add core secondary navigation.
 3. Add Progress Section format single page support.